requirements.txt
fastapi
uvicorn
requests
sentence-transformers
faiss-cpu
numpy
streamlit
pydantic